added CheckWasOpenedUseCase.kt, added test data

+++ domain/src/main/java/com/breckneck/funboxtest/domain/model/ItemDomain.kt
@@ -1,3 +1,7 @@
 package com.breckneck.funboxtest.domain.model
 
-data class ItemDomain(val id: Int,val name: String, val price: Int, val quantity: Int)
+data class ItemDomain(
+    val id: Int,
+    val name: String,
+    val price: Double,
+    val quantity: Int)
